requirements.txt
python-telegram-bot >= 12.0.0b1
requests >= 2.22.0
google-cloud-firestore